Older release of JSS had an issue with gradually slowing socket read time when transfering large files (> 10MB).  In order to test this in current and future release, there needs to be a test client that can read a file and transfer it to a server (remote or local) via JSS socket.  The server should report the number of bytes read and the time it took to read these bytes.  There should not no degradation in read time if there is no leak of any sort.
This is not part of all.pl, but is a client/server that uses JSS to transfer
files securely.  The main purpose of this test would be to test the performance
of large file transfer using JSS.

NOTE: If bufferedStream.mark(Integer.MAX_VALUE); method is invoked then fill
method of BufferedInputStream class copies lot of data using System.arraycopy
(which in-turn use memcpy). This causes very high CPU usage.  This is one of
the reasons secure large file transfer can become slow over time.