Import public key when unwrapping private key
The NSS SQL backend, unlike the DBM backend that preceded it, does
not automatically create PKCS #11 public key objects when unwrapping
private keys.  When certificates are added, this can result in a
failure to properly associate the certificate with a private key in
the token.

When unwrapping a private key, if the given public key wraps an NSS
SECKEYPublicKey object, import the public key to the token as well.

Change-Id: I146952382d535ad71d9914f3a43d7df5c0a0f510
Related: https://pagure.io/jss/issue/13