Add additional Fedora testing images
These images help ensure support for later Fedora releases and ensure
that we maintain JDK8+ support across various Fedora releases. They've
been added to the Extras section such that they don't fail the build and
aren't required to pass for the build to succeed, but so that we can
keep an eye on their status.

The fedora_29 image can be moved to the main section when 29 ships and
27 can then be removed.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>