Implement reproducible JAR builds
Because the JAR format is a zip archive, we can implement reproducible
JAR builds by extracting the zip, resetting all timestamps, and ensuring
that we add contents in sorted order. This allows us to compare built
JARs by their contents on disk versus having to extract the JAR and
compare the extracted JAR.

This adds a new make target, `reproducible` that creates a reproducible
build from a finished build, and `reproducibleCheck` that runs the test
suite against the reproducible JAR. Additionally, the spec file has been
updated to ship the reproducible JAR as well.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>