Fix DerValue.getOctetString for empty octet strings
When reading a DerValue as an OCTET STRING via getOctetString(),
if the length of the OCTET STRING is zero an exception is thrown:

    java.io.IOException: short read on DerValue buffer

The following program reproduces the issue:

    import java.io.IOException;
    import netscape.security.util.DerValue;

    class DerTest {
        public static void main(String[] args) {
            byte[] bytes = { 0x04, 0x00 };
            try {
                DerValue derVal = new DerValue(bytes);
                System.out.println(derVal.getOctetString());
            } catch (IOException e) {
                e.printStackTrace();
            }
        }
    }

The issue arises because ByteArrayInputStream.read() returns -1 when
the stream has no more data, even if we are asking to read 0 bytes.
This seems to violate the contract of InputStream.read() which says
that if the requested read length is 0, the return value is 0.
Avoid throwing the exception when this condition occurs.

Fixes: https://pagure.io/dogtagpki/issue/3079