Bump keysizes to 4096
This is required if we wish to run the JSS test suite on a system
with the FUTURE cryptopolicy currently enforcing. In particular, key
sizes of 1024 and 2048 are too small for later use by NSS, causing test
failures. Note that FUTURE currently sets a minimum key length for RSA
of 3072 bits.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>