Add JSS_FromByteArray to safely access jbyteArray
One issue with (*env)->GetByteArrayElements is that it does not NULL
terminate the input. Another issue is that forgetting to call
(*env)->ReleaseByteArrayElements results in a stray reference to the
byte array, preventing it from being garbage collected.
JSS_FromByteArray solves this by returning a new copy of the data,
releasing the reference right away.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>