Provide reverse mapping of JSSL_enums values
This is provided for convenience and to avoid rewriting the values
in-place. In certain locations, e.g., ssl/SSLSocket.c this construct is
used and the results are placed back into NSS's SSLVersionRange struct;
this is technically incorrect as the values of the struct are meant to
be NSS's SSL version constants, but are now indices into a struct.

Semantics:
 - result is the index <-> the value is found
 - result is JSSL_enums_size <-> the value is not found

This results in the possibility of the cast to a size_t if desired, but
several places in Java assume these are a (signed) int.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>