Add ssl.javax.JSSParameters
This class extends SSLParameters to provide convenience wrappers for
interacting with NSS classes in a standard Java environment. In
particular, they allow for the parsing of cipher suites into and out of
ssl.SSLCipher enum values, and for handling the differences between the
Java standard form of TLS versions as a String, ssl.SSLVersion enum
values, and ssl.SSLVersionRange instances.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>