Test SSL handshake without TCP model
The NSS documentation allows SSL_ImportFD to take a null first
parameter; in this case, NSS defaults will be used instead of defaults
from the model. Since we're creating the model *with* default parameters
(and not really saving a global model with updated parameters), don't
bother creating a model PRFileDesc.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>