Remove all usage of Assert._assert
Assert._assert is probably a relic from an older era. The class
incorrectly identifies them as "C-style assertions": the assert keyword
is actually closer to "C-style assertions" as they require the explicit
flag "-ea" (or "-enableassertions") to be passed to the JVM, similar to
how DEBUG needs to be defined for assert(...) to work in C.

In two places, where Assert._assert(false, msg) was called, I've instead
chosen to directly raise an org.mozilla.jss.util.AssertionException.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>