Introduce initial CurveDB: jss.util.ECCurve
The JDK internally maintains a database of curves (called the CurveDB);
only some of these curves are used by the Sun provider and are
internally called "named" curves. However, this curve database isn't
exposed to the public.

Unlike the JDK, we expose our curve database for other providers to use.
Currently we only maintain a mapping between names, oids, and curve
parameters. We should eventually extend it to replace all curve usages
in PK11KeyPairGenerator; this will likely involve moving the contents of
ECCurve_Code over as well.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>