Increase size of async io buffer in TestBufferPRFD
NSS on s390x seems to ignore the semantics of async io/EWOULDBLOCK:
rather than retrying the failing write, it drops the data and refuses to
continue. This manifests itself as RSA-based TLS handshakes failing,
because the certificate and associated packets exceeded the size of the
internal buffer (2048).

Bumping the buffer for the tests to 4096 plasters over the issue (so
long as we only include one certificate and not the entire chain).

redhat-bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1730109

Signed-off-by: Alexander Scheel <ascheel@redhat.com>