Implement ECPrivateKey with PK11ECPrivateKey
Since NSS exposes SECKEY_ConvertToPublicKey(...), we can convert our
private key pointer to a public key pointer and re-wrap it as a
PK11PubKey instance (of the correct subtype). This lets us implement
ECPrivateKey's getParams() interface via ECPublicKey's implementation.

Note that getS() won't be supported due to limitations in NSS. Futher
work will need to be done to expose the private key values to the caller
for this (and the interfaces for other key types) to function.
Additionally, returning the private key data might not always be
possible, e.g., when the key lives on an HSM.

This means that we're still restricted to working only with JSS-based
encryption methods for the large part.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>