Support favoring other Cryptographic Providers
We're previously given the option to either install the JSS provider or
not, and optionally to remove the Sun Provider. Users can re-order the
providers later if they want, but the most common other option will be
to only use JSS for specified operations. To achieve this, we put JSS as
the very last provider. Hence, add:

 installJSSProviderFirst = true

as a new value in InitializationValues.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>