Add KeyStoreTest back to test suite
Per report by Magnus Karlsson on pki-devel@redhat.com, KeyStoreTest is
bad example code because it doesn't actually work. In particular, it
fails to initialize a KeyStore instance correctly (passing only the name
of the provider and not the name of the KeyStore type we want).

Fix the test and re-add it to the test suite.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>