Add SSL Storage class, SSLFDProxy
Off of a SSL-enabled PRFileDesc created by NSS's SSL_ImportFD, we might
want to store various information which would be helpful (and, necessary
to free at the end):

 - Any TrustManagers we're using on this connection,
 - Whether or not the handshake has finished,
 - The client certificate, or
 - Additional parameters required for NSS callbacks.

Some of these will be Java-backed values, in which case they should be
added directly to SSLFDProxy. Others will be C-backed values, in which
case they should be freed once the SSL Socket is closed.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>