Revert "Fix build order: tests depend on a built JAR"
This reverts commit a52de916344cfc6f67b8b1178699ca5b6b559d41. This
commit isn't necessary: while the tests depend on the rest of JSS, javac
will detect the necessary files (because of -sourcepath) and continue
with the build.

In fact, build order is a red herring: the real issue is that the
org.mozilla.jss.pkix package namespace is empty, so:

    import org.mozilla.jss.pkix.*;

won't have any classes to import.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>