Wrap a certificate with chain to jobjectArray
Sometimes we have only a single CERTCertificate but wish to inquire
about as much of its chain as we can, and return the result as a
jobjectArray to pass back to Java. This is helpful when we've gotten a
CERTCertificate from NSS and we need to check it against a TrustManager
instance.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>