Capture PRErrorCode <-> Exception mappings
Often we'll need to take a PRErrorCode and raise it as a Java Exception
(via the JNI interface) of the proper type. Other times, we'll need to
take a raised Java Exception and return it as the correct
(SECStatus, PRErrorCode) pair.

We introduce JSS_ExceptionToSECStatus and JSS_SECStatusToException to
handle this. The latter has a *Message form, which takes an error
message to raise with the exception.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>