Support getting the current OCSP policy enum
By exposing the value of ocspPolicy from the CryptoManager (instead
of the ordinal, which we'd have to convert back to an enum for most
calls), we can more easily support save/restore operations for the value
of OCSP policy.

This will enable, in combination with the previous patch, explicit OCSP
checks on certificates and their chains, restoring the current policy
afterwards.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>