Support using SymmetricKey in JSSMacSpi
JSSMacSpi differs from JSSCipherSpi in that it requires the caller to
pass a SecretKeyFacade instance (which merely wraps a SymmetricKey) and
are typically created by SecretKeyFactory. However, the caller might
have an existing SymmetricKey and wish to use that with JSSMacSpi; check
for that case and handle it gracefully. Unlike JSSCipherSpi, we don't
use the SecretKeyFactory to clone the underlying key.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>