Enable SSLAlertEvent storage on SSLFDProxy
Storing events on a SSLFDProxy object requires:

 - A place to put them (both inbound and outbound),
 - A way of tracking how many we've processed,
 - A way of giving the SSLFDProxy reference to NSS (a GlobalRefProxy).

Unlike the client certificate callback (where we can directly pass the
CERTCertificate pointer and assume it is valid for the entire lifetime
of the SSLFDProxy), we have to have this global state in order to
interact with Java-owned objects.

We introduce helper functions for accessing and creating all
event-related members and add the underlying callbacks one would use to
trigger them. In the future, other callbacks could be written that
trigger an event (similar to the existing SSLSocket code); this isn't
suitable for use in an SSLEngine however.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>