Make NativeProxy AutoCloseable
We extend the default NativeProxy implementation in three directions:

 - Make it AutoCloseable,
 - Add a clear() method which supports clearing the value of the
   pointer without calling releaseNativeResources(...), and
 - Save a stack trace when assertions are enabled.

By implementing AutoCloseable, we enable the common Java pattern of
try-with-resources:

    try (NativeProxy item = ...) {
        ...
    }

and have the item automatically be closed (in our case, finalized(...))
at the end of its life cycle. We've added additional logic to prevent
double-free bugs.

Additionally, by adding clear(...), we both add protection from
double-free bugs and enable resource cleanup from native handlers that
removes entries in NativeProxy's internal tracking.

Lastly, we save stack traces when assertions are enabled in the JVM and
JSS is built in DEBUG mode (with DEBUG defined).

Signed-off-by: Alexander Scheel <ascheel@redhat.com>