Introduce GlobalRefProxy
Java JNI has the notion of a global reference to an object, instead of
the traditional local reference created by the JVM when calling any
JNI'd function. Since all global references have the same type (jobject)
and release handler ((*env)->DeleteGlobalRef(...)), it makes sense to
track these as a single JNI type.

This ensures that these resources are released, allowing the referenced
object to be garbage collected.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>