Make FindNSS and FindNSPR return directories
We pass NSS_LIBRARIES and NSPR_LIBRARIES to the compiler via -L (capital
means directory to search for libraries). However, FindNSS and FindNSPR
returned the paths to the actual libraries in these variables. Make them
return the directories these libraries reside in instead, minimizing the
list if possible.

This means that we'll now pass the extraneous -L/usr/lib64 in most
cases, but in the case of sandbox builds, this will be correct. Further,
it enables the usage of these arrays on LD_LIBRARY_PATH in a future
commit.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>