Add {NSS,NSPR}_LIBRARIES to LD_LIBRARY_PATH
By adding NSS_LIBRARIES and NSPR_LIBRARIES to LD_LIBRARY_PATH, builds of
JSS using a sandboxed NSS build will be able to find other libraries
required by NSS (such as NSS modules). This is especially a problem when
the version of NSS on the system greatly differs from that of the
sandbox build.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>