Move javac command to file lists
javac supports parsing options from a file using the @/path/to/arguments
syntax. This lets us pass all of our sources to javac in a single file
(containing all of the Java files we wish to build in this pass) instead
of passing them individually on the command line. This reduces the size
of the output significantly when there's an error in the build.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>