Remove code dependency on Apache Commons Codec
Since Base64 is included in the standard library since JDK 8, the
minimum JDK version JSS master supports, and we already have hex
encoding in netscape.security.utils.Utils, we can drop all of our
dependencies on Apache Commons Codec fairly easily. Doing so reduces the
footprint of the JSS library and otherwise is beneficial since there's
likely to be little need for it in the future.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>