Introduce NativeEnclosure and StaticVoidPointer
A NativeEnclosure is an abstract class which stores a NativeProxy
instance and wraps both the pointer and the size of the memory pointed
to by said pointer. When used in combination with StaticVoidPointer, a
pointer which isn't tracked and freed like the usual NativeProxy is,
this can be used as a form of serialization, especially for converting
parameters to and from the JNI layer.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>