Implement KBKDF Parameters
Underlying all three KBKDFs are a common set of parameters represented
in the base KBKDFParameterSpec class: information about the underlying
PRF, information about the input to the PRF, and information about the
output of the PRF.

The input to the PRF are a series of KBKDFDataParameters, which there
are four types of:

 1. KBKDFByteArrayParam, a fixed byte array which gets passed to the
    KDF,
 2. KBKDFDKMLengthParam, an indication of how to compute and encode the
    length of the PRF output stream,
 3. KBKDFIterationVariableParam, an indication of where to place the
    iteration information (whether it is the required counter in Counter
    Mode or the previous PRF's output in Feedback and Double Pipeline
    modes),
 4. KBKDFOptionalCounterParam, an indication of how to encode an
    optional counter in Feedback and Double Pipeline modes.

These parameters are index sensitive: they're added to the PRF input
stream in the order they're specified.

There's also a series of derived keys. First is the implicit derived
key, which gets returned by the call to generateKey() -- however, other
derived keys can get specified by adding KBKDFDerivedKey objects to the
KBKDF parameters. No two keys may overlap in the same PRF segment.

Note that this commit only adds the data parameters and the required
NativeEnclosure serializations; no KDF operations are performed.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>