Enable SSLAlertEvent to raise SSLExceptions
When using SSLAlertEvents inside an SSLEngine, we need to raise a
SSLException of the correct type. There's one issue here though:

While we could add support for telling if this is an inbound or
an outbound alert, we don't know _what_ caused said alert (because
NSS created it) -- an issue on the client or an issue on the server.
This means that we can't report a SSLKeyException and only report
SSLProtocolExceptions.

Considering NSS is the one sending the alert, and we're just exposing
it happened to the calling SSLEngine, this is a fine trade-off.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>