Fix handling of global_ref_proxy
In JSS, a GlobalRefProxy stores a JNI global jobject reference, to allow
Java to clean it up later. This allows us to pass the ref to a NSS SSL
callback, assuming we promise to clean it up later (and not clean it up
before NSS is done with it!).

This fixes various issues with the handling of global references:

 - We needn't clear the value on successful storage, as we intend to
   continue using said reference.
 - We need to validate that the ref isn't NULL, when created.
 - We need to validate JSS_javaVM is non-NULL (so we can extract a
   JNI ENV reference).

Signed-off-by: Alexander Scheel <ascheel@redhat.com>