Clear errors before executing most NSPR calls
NSPR behaves much like the C standard library's errno; unless the value
is cleared, it is hard to tell if an exception occurred in this call.
Because these are relatively unopinionated bindings over NSPR's calls,
clear the error message rather than making the caller do so (and incur
another JNI penalty).

Signed-off-by: Alexander Scheel <ascheel@redhat.com>