Fix memory leaks around Array handling
Reported by @c42-arthur in #371.

PK11Cipher's updateContext leaks a reference to the input byte array, a
potentially highly-used method depending on the application. This fixes
that memory leak and audits the rest of JSS for similar leaks, fixing
two in the NSPR and NSS SSL bindings.

The case in ssl/javasock.c is at worst a memcpy of a NULL variable, which
should fail where it to be hit.

Resolves: #371

Signed-off-by: Alexander Scheel <ascheel@redhat.com>