Workaround NSS PR_Read bug
When using SSL on certain systems, NSS will indicate it only has a
single byte of data to read, when in fact it has more data than that.
This is due to a bug in the TLS record parsing logic, forcing the caller
to issue another PR_Read call.

This hangs up our SSLEngine implementation. Bypass this bug by detecting
if the underlying socket is non-blocking, and if so, call PR_Read until
we get a PR_WOULD_BLOCK_ERROR.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>