Advanced SSLCipher feature detection
Mimic the behavior in NSS's ssl3_config_match_init to ensure we limit
ourselves to ciphers that NSS will actually negotiate. In particular,
validate that there exists a token that could perform the given
authentication, hash, or key type if it was required of us.

This largely mimics code in NSS's ssl3con.c.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>