Add JSSProviderLoader implementation
Similar to the way the SunPKCS11 provider works, expose an alternative
JSSProvider, JSSProviderLoader, which loads configuration from a path or
InputStream, and configures our CryptoManager correctly. This allows us
to inject the provider via the java.security interface, rather than via
the existing CryptoManager.initialize(...) method.

Currently this only supports using a single, fixed password. In the
future, JSSProvider could be converted into an AuthProvider (which
supports logging-in to arbitrary tokens), or a more comprehensive
File-based PasswordCallback implementation added.

Additionally, this provider (unlike the JSSProvider it derives from)
supports the JDK9+ style of static initialization, accepting a path to a
configuration file.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>