Convert test suite to JSSProviderLoader
This converts the test suite from the previous
CryptoManager.initialize() based approach to the new java.security
approach. This should work on most platforms. If this fails on your
specified platform, it is likely due to the providers contained in
tools/java.security.in; make sure to update them to match your system's
file.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>