Add detection for SSLCipherInfo's peerDelegCred
Trust but verify. NSS usually documents the version a feature was added
in. However, this wasn't documented in the peerDelegCred case; because
the comment above referred to the following fields being added in NSS
3.34, I assumed it held for this field as well. However, according to
the commit history, it was added more recently, in NSS v3.45.

Add feature detection and another conditional field for this.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>