SSLSocket: detect failure to recieve CLOSE_NOTIFY
An important part of the TLS protocol is the ability to detect whether
or not the peer closed the connection or if an attacker terminated the
connection prematurely. In order to do this, both sides send a
CLOSE_NOTIFY event. When this event isn't detected, raise an IOException
to inform the application about it.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>