Introduce ByteBuffer-optimized buffer for PRI/O
j_bytebuffer wraps the byte array underlying a ByteBuffer; this allows
us to call PR.Read and have it read directly from the underlying
ByteBuffer, rather than copying it into a temporary buffer before the
Read call. This, in turn, means we're making fewer JNI round-trips: we
only need to pass the array to the underlying j_bytebuffer, call
PR_Read, and it'll handle getting all the data it can in a single pass,
resulting in fewer loops inside JSSEngine.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>