Simplify JSSServerSocketChannel.accept()
Since we won't return the JSSServerSocketChannel in situations where the
wrapped socket lacks a channel, it is highly unlikely that this method
will ever be called when the parent channel is null. Simplify the logic
and throw an exception in this case instead of returning an accepted
socket channel. Note that, while calling accept() on the parent socket will
allows return a non-null socket, it isn't guaranteed to have a
SocketChannel.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>