Fix race in configuring session cache
When multiple JSSSockets and/or JSSEngines are started concurrently,
they will race to configure the session cache. However, the cache can
only be created by one thread at a time and only needs to be done once
in the lifetime of JSS. Make JSSEngine take ownership of cache creation
and utilize this within the legacy SSLSocket as well.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>