Support asynchronous certificate authentication
This exposes support for the SSL_AuthCertificateComplete API call, which
allows us to create a SSL_AuthCertificateHook handler which returns
SECWouldBlock. This returns control back to the caller, allowing them to
execute certificate authentication on their own, lazily. When the result
is available, the Complete call can be performed with the result of the
check. This will enable us to do external (non-native) X509TrustManager
checking in JSS's SSLEngine implementation in a Runnable, implementing
the DelegatedTask aspect.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>