Handle NSS/TLS SSLVersion identifiers
NSS uses the TLS protocol version identifiers from the specs (two byte
integers, 0x0002 for SSLv2, 0x0300 for SSLv3, and 0x030{version+1} for
TLSv1.0 -> TLSv1.3); when the caller forgets to convert the value to the
internal JSS enum index, they're left holding a NULL enum value.
Instead, track the real NSS value and return the proper Enum for it.
This shouldn't be a problem as these values are mostly unique (sans a
SSLv2 / SSL_ENABLE_TLS conflict with value 0x0002.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>