Use JSS-provided TrustManagers by default
When SSLContext is called with no TrustManagers, we end up defaulting to
the default internal NSS validation logic. This results in connections
which fail to verify when no hostname is provided, because NSS strictly
requires hostnames to be present and matching. Because the JDK otherwise
requires hostnames to validated (by e.g., a HostnameVerifier), we
shouldn't use the NSS handler by default.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>