Duplicate client certificate in handler
When JSS passes the certificate for use in the client auth handler, it
doesn't duplicate this certificate. However, NSS will later attempt to
free this key. We should duplicate the key before returning it to NSS,
allowing NSS to free it safely.

Note that, because the key isn't passed in to the client auth handler,
but instead queried, we need not duplicate it.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>