Add Post-Handshake Auth to legacy SSLSocket
When TLSv1.3 was introduced into the legacy SSLSocket prior to the
introduction of JSSEngine, we didn't add support for PHA. This only
requires adding a single option, SSL_ENABLE_POST_HANDSHAKE_AUTH, and
allowing the user to control whether or not it is enabled. Enabling this
option brings compatibility with TLSv1.2's rehandshakes for identifying
the user after the initial request. Because Tomcat uses this upgrade
path (for an application with mixed authenticated and non-authenticated
endpoints), we need to support it in SSLSocket.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>